Module Main.Make

Parameters

module Conf : Sc_lib.Config.T

module Database : Sc_lib.Tool.Database

Signature

module ToolInfos : sig ... end

include module type of struct include ToolInfos end

include module type of struct include TomlSpec end

type opt = TomlSpec.opt = {

custom_tests : string list;

}

val toml_schema : opt Sc_lib.Eztoml.schema

val default_opt : opt

val tool_kind : [> `Preprocess ]

include sig ... end

val doc : string

val module_name : string

val tool_name : string

module Log : sig ... end

module Log_lwt : sig ... end

val tool_project_dir : Sc_sys.File.TYPES.dir

type label_family = ToolInfos.label_family

type input_data = Database.Runtime.working_data

type working_data = ToolInfos.working_data

val output_file : ?funname:string -> input_data -> [ `C ] Sc_sys.File.TYPES.file

module Lreplay : sig ... end

type runtime_data = Database.Runtime.runtime_data

val map_on_option_testfiles : opt:opt -> ([> `C ] Sc_sys.File.t -> [> `C ] as 'a Sc_sys.File.t) -> 'b Sc_sys.File.t list

val is_installed : unit -> bool Lwt.t

val imported_tests_dir : Sc_sys.File.dir

val existing_files_digests : Stdlib.Digest.t list Lwt.t

Lists the digest files already imported into the replayer directory.

val replayer_main_o : [ `O ] Sc_sys.File.t Lwt.t

val test_already_registered : [> `C ] Sc_sys.File.t -> bool Lwt.t

Compares the digests of the files in the test dir with the one in argument.

val write_test_with_extra_header : dir:Sc_sys.File.dir -> labelized_file:[> `labelized ] Sc_sys.File.t -> [> `C ] Sc_sys.File.t -> 'a Sc_sys__File.TYPES.file

Adds to a test file the replayer header driver, initializing pc_label macros.

val preprocess_file : opt:opt -> Database.Runtime.working_data -> {working_data}4 Lwt.t

Copies tests given by the user into the working directory.

val no_preprocess_file : opt:opt -> Database.Runtime.working_data -> {working_data}4 Lwt.t

val intermediate_main : string

type testcase_workspace = {

dir : Sc_sys.File.dir;

testcase : [ `C ] Sc_sys.File.t;

}

val init_testcase_workspace : labelized_file:[> `labelized ] Sc_sys.File.t -> [ `C ] Sc_sys.File.t -> testcase_workspace Lwt.t

Creates a testcase workspace.

val header_dirs_with_i : string list

The list of header directories to give to clang.

val play : shareddb:Database.Runtime.runtime_data -> testcase_workspace -> bool Lwt.t

Builds an executable program corresponding to the given custom test, and executes it. Updates the shared status store unless an RTE occurs before the program terminates.

val copy_test_in_testcase_dir : labelized_file:[ `C | `labelized ] Sc_sys.File.t -> 'a Sc_sys__File.TYPES.file -> unit

Writes a custom test in the main testcases directory

val generate_test_cases : opt:'a -> shareddb:Database.Runtime.runtime_data -> working_data -> 'b -> unit Sc_sys.Process.t Lwt.t